# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todos API",
        "description": "Dokumentasi API Todos",
        "contact": {
            "email": "rikinofianto@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/todo/add": {
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "Membuat todo baru",
                "operationId": "690124e67f04be53b43060f4b9f8b151",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Mengerjakan task CRUD API"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "waktu 24 jam"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data sukses dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Todo created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/todo/list": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Mendapatkan list data todos user",
                "operationId": "0737e01aab8aea4572bf9aefa41de50d",
                "responses": {
                    "200": {
                        "description": "Data sukses diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "description": "Daftar objek todo",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Todo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/todo/{id}": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Mendapatkan satu data todos user",
                "operationId": "830c0c2045aff089c27e9a8bbe3de8cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID todo yang ingin diambil",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data sukses diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "description": "Object data todo",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/todo/update/{id}": {
            "put": {
                "tags": [
                    "Todos"
                ],
                "summary": "Mengubah data todos user",
                "operationId": "4f66d46fa60e30f412080ef7f93496b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID todo yang ingin diubah",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data sukses diubah",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Todo updated successfully"
                                        },
                                        "data": {
                                            "description": "Object data todo",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/todo/delete/{id}": {
            "delete": {
                "tags": [
                    "Todos"
                ],
                "summary": "Menghapus data todos user",
                "operationId": "b6d12badd8f39b6a9fc0b7b5b8d67abe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID todo yang ingin dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data sukses dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Todo deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registrasi Pengguna",
                "operationId": "d25ccfe75d772272fa3e5adbe0636701",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "bejo dikaryo"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registrasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User registered successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login Pengguna dan Mendapatkan Token JWT",
                "operationId": "df2c1d79263b8ad0f88e3f04c5767310",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Kredensial tidak valid"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Todo": {
                "title": "Todo Model",
                "description": "Todo data structure",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "title": {
                        "type": "string",
                        "example": "Makan"
                    },
                    "description": {
                        "type": "string",
                        "format": "text",
                        "example": "pake telur"
                    },
                    "is_done": {
                        "type": "string",
                        "format": "int",
                        "example": 0
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT Anda di sini (tanpa 'Bearer ')",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Todos",
            "description": "Todos"
        },
        {
            "name": "Authentication",
            "description": "Authentication"
        }
    ]
}